add insulin tracking: update userInsulinOnBoard type

+++ Code/machine.h
@@ -78,8 +78,8 @@ private:
     int currentPage;
     double currentInsulinAmount; // This is the amount of insulin in the machine in u, will be out of 300 u
     int currentBatteryLevel;
-    int userGlucoseLevel;   // Current Blood Glucose of User, in mmol/L
-    int userInsulinOnBoard; // The amount of insulin still inside the body after a bolus injection
+    int userGlucoseLevel;      // Current Blood Glucose of User, in mmol/L
+    double userInsulinOnBoard; // The amount of insulin still inside the body after a bolus injection
 
     bool isCharging;
     bool isTurnedOn;
